move and rename bootstrap overrides
