+++ src/bricks/renderers/documentView/DocumentView.tsx
@@ -19,7 +19,7 @@ import { buildDocumentBranch } from "@/components/documentBuilder/documentTree";
 import React from "react";
 import EmotionShadowRoot from "react-shadow/emotion";
 import bootstrap from "bootstrap/dist/css/bootstrap.min.css?loadAsUrl";
-import bootstrapOverrides from "@/bootstrapOverrides.scss?loadAsUrl";
+import bootstrapOverrides from "@/pageEditor/sidebar/sidebarBootstrapOverrides.scss?loadAsUrl";
 
 import { type DocumentViewProps } from "./DocumentViewProps";
 import DocumentContext from "@/components/documentBuilder/render/DocumentContext";
